[#17] Refactor: params가 없는 페이지들 router 폴더로 이동
